move gfx out of slangrt
